:recycle: Refactor render structures
