+++ render-wasm/src/shapes/modifiers/grid_layout.rs
@@ -315,8 +315,6 @@ fn set_flex_multi_span(
         for track in tracks[start..end].iter_mut() {
             if track.track_type == GridTrackType::Flex {
                 let new_size = alloc.clamp(track.size, track.max_size);
-                let aloc = new_size - track.size;
-                dist -= aloc;
                 track.size = new_size;
             }
         }
